move deep merge and time cache into separate modules
