+++ paasta_tools/kafkacluster_tools.py
@@ -19,8 +19,8 @@ import service_configuration_lib
 from paasta_tools.kubernetes_tools import sanitised_cr_name
 from paasta_tools.long_running_service_tools import LongRunningServiceConfig
 from paasta_tools.long_running_service_tools import LongRunningServiceConfigDict
+from paasta_tools.util.deep_merge import deep_merge_dictionaries
 from paasta_tools.utils import BranchDictV2
-from paasta_tools.utils import deep_merge_dictionaries
 from paasta_tools.utils import DEFAULT_SOA_DIR
 from paasta_tools.utils import load_service_instance_config
 from paasta_tools.utils import load_v2_deployments_json
