+++ paasta_tools/paasta_service_config_loader.py
@@ -21,7 +21,7 @@ from typing import Type
 from service_configuration_lib import read_service_configuration
 
 from paasta_tools import utils
-from paasta_tools.utils import deep_merge_dictionaries
+from paasta_tools.util.deep_merge import deep_merge_dictionaries
 from paasta_tools.utils import DEFAULT_SOA_DIR
 from paasta_tools.utils import InstanceConfig_T
 from paasta_tools.utils import list_clusters
